Decouple Opengl settings from world
